+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-## 1.0.2
-修复QuadTree find方法
+## 1.0.5
+ TreeNode add computeDeep method
+
+## 1.0.4
+fix QuadTree find method
 
 * TODO: Describe initial release.